refactor(authorization): use module-tools chore(authorization): optimize imports & cleanup code
