+++ modules/authorization/src/Authorization.ts
@@ -1,10 +1,8 @@
 import ConduitGrpcSdk, {
-  ConfigController,
   DatabaseProvider,
   GrpcRequest,
   GrpcResponse,
   HealthCheckStatus,
-  ManagedModule,
 } from '@conduitplatform/grpc-sdk';
 import path from 'path';
 import AppConfigSchema, { Config } from './config';
@@ -29,10 +27,10 @@ import { RelationsController } from './controllers/relations.controller';
 import { ResourceController } from './controllers/resource.controller';
 import { AdminHandlers } from './admin';
 import { status } from '@grpc/grpc-js';
+import { ConfigController, ManagedModule } from '@conduitplatform/module-tools';
 
 export default class Authorization extends ManagedModule<Config> {
   configSchema = AppConfigSchema;
-  protected metricsSchema = metricsSchema;
   service = {
     protoPath: path.resolve(__dirname, 'authorization.proto'),
     protoDescription: 'authorization.Authorization',
@@ -50,6 +48,7 @@ export default class Authorization extends ManagedModule<Config> {
       can: this.can.bind(this),
     },
   };
+  protected metricsSchema = metricsSchema;
   private adminRouter: AdminHandlers;
   private indexController: IndexController;
   private permissionsController: PermissionsController;
@@ -66,16 +65,6 @@ export default class Authorization extends ManagedModule<Config> {
     await this.grpcSdk.waitForExistence('database');
     this.database = this.grpcSdk.database!;
     await runMigrations(this.grpcSdk);
-  }
-
-  protected registerSchemas() {
-    const promises = Object.values(models).map(model => {
-      const modelInstance = model.getInstance(this.database);
-      return this.database
-        .createSchemaFromAdapter(modelInstance)
-        .then(() => this.database.migrate(modelInstance.name));
-    });
-    return Promise.all(promises);
   }
 
   async onConfig() {
@@ -204,6 +193,7 @@ export default class Authorization extends ManagedModule<Config> {
     //   this.grpcSdk.registerMetric(metric.type, metric.config);
     // }
   }
+
   createResourceObject(
     name: string,
     relations: Resource_Relation[],
@@ -225,5 +215,15 @@ export default class Authorization extends ManagedModule<Config> {
       resource.permissions![permission.name] = permission.roles;
     });
     return resource;
+  }
+
+  protected registerSchemas() {
+    const promises = Object.values(models).map(model => {
+      const modelInstance = model.getInstance(this.database);
+      return this.database
+        .createSchemaFromAdapter(modelInstance)
+        .then(() => this.database.migrate(modelInstance.name));
+    });
+    return Promise.all(promises);
   }
 }
