+++ modules/authorization/src/controllers/index.ts
@@ -1,0 +1,5 @@
+export * from './cache.controller';
+export * from './index.controller';
+export * from './permissions.controller';
+export * from './relations.controller';
+export * from './resource.controller';
