+++ modules/authorization/src/models/ObjectIndex.schema.ts
@@ -1,4 +1,5 @@
-import { ConduitActiveSchema, DatabaseProvider, TYPE } from '@conduitplatform/grpc-sdk';
+import { DatabaseProvider, TYPE } from '@conduitplatform/grpc-sdk';
+import { ConduitActiveSchema } from '@conduitplatform/module-tools';
 
 /**
  * This is used to model a relation tuple like:
